Risolti i problemi riscontrati dopo il precedente commit per cui ricarico  i file per il nuovo sistema di backup di GAzie: Uusando la nuova classe MySQL-dump, Ripristinato il corretto funzionamento dei backup del data base di GAzie, in automatico e in manuale. NB: i file di backup adesso hanno l'estensione .gz. Tale estensione viene direttamente accettata da Phpmyadmin. Per aprirli manualmente è necessario decomprimerli usando appositi programmi, ad esempio 7-zip
git-svn-id: https://svn.code.sf.net/p/gazie/code/trunk@6425 0ed6177e-08db-47c9-9b37-4a62b3198fc7

+++ modules/inform/report_backup.php
@@ -26,12 +26,41 @@ require("../../library/include/datlib.inc.php");
 $admin_aziend = checkAdmin(9);
 require("../root/lib.function.php");
 $checkUpd = new CheckDbAlign;
+$kb=0;
+?>
+<style>
+	#loader {
+		border: 12px solid #f3f3f3;
+		border-radius: 50%;
+		border-top: 12px solid #444444;
+		width: 70px;
+		height: 70px;
+		animation: spin 1s linear infinite;
+	}
+
+	@keyframes spin {
+		100% {
+			transform: rotate(360deg);
+		}
+	}
+
+	.center {
+		position: absolute;
+		top: 0;
+		bottom: 0;
+		left: 0;
+		right: 0;
+		margin: auto;
+	}
+</style>
+<div id="loader" class="center"></div>
+<?php
+
 //
 // Verifica i parametri della chiamata.
 //
 require("../../library/include/header.php");
 $script_transl = HeadMain();
-
 if (isset($_POST['hidden_req'])) { // accessi successivi allo script
     $form['hidden_req'] = $_POST["hidden_req"];
     $form['ritorno'] = $_POST['ritorno'];
@@ -40,7 +69,8 @@ if (isset($_POST['hidden_req'])) { // accessi successivi allo script
         $nv = filter_input(INPUT_POST, 'backup_mode');
         $kb = filter_input(INPUT_POST, 'keep_backup');
         $fs = filter_input(INPUT_POST, 'freespace_backup');
-        $fi = filter_input(INPUT_POST, 'filebackup');
+        //$fi = filter_input(INPUT_POST, 'filebackup');
+        $fi = (isset($fi))?$fi:'0';
         $checkUpd->backupMode($nv); // passando un valore alla stessa funzione faccio l'update
         gaz_dbi_put_row($gTables['config'], 'variable', 'keep_backup', 'cvalue', $kb);
         gaz_dbi_put_row($gTables['config'], 'variable', 'freespace_backup', 'cvalue', $fs);
@@ -54,7 +84,7 @@ if (isset($_POST['hidden_req'])) { // accessi successivi allo script
 $bm = $checkUpd->backupMode();
 $keep = gaz_dbi_get_row($gTables['config'], 'variable', 'keep_backup');
 $freespace = gaz_dbi_get_row($gTables['config'], 'variable', 'freespace_backup');
-$filebackup = gaz_dbi_get_row($gTables['config'], 'variable', 'file_backup');
+//$filebackup = gaz_dbi_get_row($gTables['config'], 'variable', 'file_backup');
 ?>
 
 <form method="POST">
@@ -80,7 +110,7 @@ $filebackup = gaz_dbi_get_row($gTables['config'], 'variable', 'file_backup');
                     <th class="FacetFieldCaptionTD"><?php echo $script_transl['id']; ?></th>
                     <th class="FacetFieldCaptionTD"><?php echo $script_transl['ver']; ?></th>
                     <th class="FacetFieldCaptionTD"><?php echo $script_transl['name']; ?></th>
-                    <th class="FacetFieldCaptionTD"><?php echo $script_transl['size']; ?></th>            
+                    <th class="FacetFieldCaptionTD"><?php echo $script_transl['size']; ?></th>
                     <th class="FacetFieldCaptionTD"><?php echo $script_transl['dow']; ?></th>
                     <?php
                         if ( $admin_aziend["Abilit"]=="9") {
@@ -94,7 +124,7 @@ $filebackup = gaz_dbi_get_row($gTables['config'], 'variable', 'file_backup');
                 $files = array();
                 if ($handle = opendir(DATA_DIR.'files/backups/')) {
                     while (false !== ($file = readdir($handle))) {
-                        if ($file != "." && $file != ".." && strpos($file, ".gaz")) {
+                        if ($file != "." && $file != ".." && strpos($file, ".gz")) {
                             $files[filemtime(DATA_DIR.'files/backups/' . $file)] = $file;
                         }
                     }
@@ -104,15 +134,16 @@ $filebackup = gaz_dbi_get_row($gTables['config'], 'variable', 'file_backup');
                     $index = 0;
                     $id = array ();
                     foreach ($files as $file) {
+
                         preg_match('/-(.*?)-/',$file, $id);
-                        
-                        if ($index < 30) {
+
+                        if ($index < 30) { // se il modo è manuale visualizzo solo gli ultimi 30 backup, gli eccedenti li cancello
                             ?>
-                            <tr class="FacetDataTD"><td><a class="btn btn-xs btn-default" href="">
+                            <tr class="FacetDataTD"><td><a class="btn btn-xs btn-default" style="cursor:default;" href="">
                                         <?php echo (count($id)>0) ? $id[1] : "nd"; ?>
                                     </a></td>
                                 <td>
-                                    <?php 
+                                    <?php
                                         if ( preg_match('/-v(.*?).sql/',$file, $versione)>0 )
                                             echo $versione[1];
                                     ?>
@@ -138,8 +169,8 @@ $filebackup = gaz_dbi_get_row($gTables['config'], 'variable', 'file_backup');
                             </tr>
                             <?php
                             $index++;
-                        } else {
-                            unlink(DATA_DIR."files/backups/" . $file);
+                        } elseif ( $bm !== "automatic"){
+                           unlink(DATA_DIR."files/backups/" . $file);
                         }
                     }
                 }
@@ -171,7 +202,7 @@ $filebackup = gaz_dbi_get_row($gTables['config'], 'variable', 'file_backup');
                     <div class="div-table-col">
                         <h4>Automatico</h4>
                         <ul class="licheck">
-                            <li>Esegui backup dei files (la cartella di gazie verrà salvata) : <input <?php echo ($filebackup['cvalue']==1) ? 'checked="checked"' : ''; ?> type="checkbox" name="filebackup" value="1" /> </li>
+                           <!-- <li>Esegui backup dei files (la cartella di gazie verrà salvata) : <input <?php echo ($filebackup['cvalue']==1) ? 'checked="checked"' : ''; ?> type="checkbox" name="filebackup" value="1" /> </li> -->
                             <li>Numero di backup da conservare : <input type="text" name="keep_backup" value="<?php echo $keep['cvalue']; ?>" /> inserire 0 per tutti</li>
                             <li>Spazio da lasciare libero (%) : <input type="text" name="freespace_backup" value="<?php echo $freespace['cvalue']; ?>" /> raggiunto il limite i backup vecchi verranno cancellati</li>
                         </ul>
@@ -187,6 +218,17 @@ $filebackup = gaz_dbi_get_row($gTables['config'], 'variable', 'file_backup');
     </div>
 
 </form>
+<script>
+document.onreadystatechange = function() {
+    if (document.readyState !== "complete") {
+        document.querySelector("body").style.visibility = "hidden";
+        document.querySelector("#loader").style.visibility = "visible";
+    } else {
+        document.querySelector("#loader").style.display = "none";
+        document.querySelector("body").style.visibility = "visible";
+    }
+};
+</script>
 <?php
 require("../../library/include/footer.php");
 ?>
